as,gs,is,js,ns,util: Move `pkg/telemetry` to `pkg/telemetry/tracing`
