+++ pkg/config/shared.go
@@ -32,7 +32,7 @@ import (
 	"go.thethings.network/lorawan-stack/v3/pkg/httpclient"
 	"go.thethings.network/lorawan-stack/v3/pkg/log"
 	"go.thethings.network/lorawan-stack/v3/pkg/redis"
-	"go.thethings.network/lorawan-stack/v3/pkg/telemetry"
+	"go.thethings.network/lorawan-stack/v3/pkg/telemetry/tracing"
 	"gocloud.dev/blob"
 )
 
@@ -505,7 +505,7 @@ type ServiceBase struct {
 	Rights           Rights               `name:"rights"`
 	KeyVault         KeyVault             `name:"key-vault"`
 	RateLimiting     RateLimiting         `name:"rate-limiting" description:"Rate limiting configuration"`
-	Telemetry        telemetry.Config     `name:"telemetry" yaml:"telemetry" description:"Telemetry configuration"`
+	Tracing          tracing.Config       `name:"tracing" yaml:"tracing" description:"Tracing configuration"`
 	SkipVersionCheck bool                 `name:"skip-version-check" yaml:"skip-version-check" description:"Skip version checks"` //nolint:lll
 }
