+++ pkg/telemetry/tracing/config.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package telemetry
+package tracing
 
 import "go.thethings.network/lorawan-stack/v3/pkg/config/tlsconfig"
